Fixing 'last' of RECONR -> JSON issues.
