+++ src/lipservice/RECONR/Card5/test/Card5.test.cpp
@@ -12,8 +12,8 @@ SCENARIO( "bugless" ){
   CHECK( value == card5.cards.value );
 
   THEN( "Card5 can be turned to JSON" ){
-    nlohmann::json refJSON = { { "cards", value } };
-    CHECK( refJSON == card5 );
+    nlohmann::json refJSON = value;
+    CHECK( refJSON == nlohmann::json( card5 ) );
   } // THEN
 }
